requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pydantic>=2.9,<3
numpy>=2.3,<3
faiss-cpu>=1.13,<2
PyMuPDF>=1.26,<2
requests>=2.32,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<1
